+++ src/oscar/Maths/PolarPerspectiveCamera.h
@@ -65,6 +65,7 @@ namespace osc
     PolarPerspectiveCamera CreateCameraWithRadius(float);
     PolarPerspectiveCamera CreateCameraFocusedOn(AABB const&);
     Vec3 RecommendedLightDirection(PolarPerspectiveCamera const&);
+    void FocusAlongAxis(PolarPerspectiveCamera&, size_t, bool negate = false);
     void FocusAlongX(PolarPerspectiveCamera&);
     void FocusAlongMinusX(PolarPerspectiveCamera&);
     void FocusAlongY(PolarPerspectiveCamera&);